Audio samplerates can get past a short capacity, 48kHz tested to be working on i6c

+++ src/hal/hisi/v3_hal.c
@@ -72,7 +72,7 @@ void v3_audio_deinit(void)
     v3_aud.fnDisableDevice(_v3_aud_dev);
 }
 
-int v3_audio_init(short samplerate)
+int v3_audio_init(int samplerate)
 {
     int ret;
