+++ src/hal/star/i6_hal.h
@@ -22,7 +22,7 @@ int i6_hal_init(void);
 void *i6_audio_thread(void);
 
 void i6_audio_deinit(void);
-int i6_audio_init(short samplerate);
+int i6_audio_init(int samplerate);
 
 int i6_channel_bind(char index, char framerate);
 int i6_channel_create(char index, short width, short height, char mirror, char flip, char jpeg);
