+++ src/hal/star/i6c_hal.c
@@ -75,7 +75,7 @@ void i6c_audio_deinit(void)
     i6c_aud.fnDisableDevice(_i6c_aud_dev);
 }
 
-int i6c_audio_init(short samplerate)
+int i6c_audio_init(int samplerate)
 {
     int ret;
