+++ src/hal/star/i6c_hal.h
@@ -19,7 +19,7 @@ void i6c_hal_deinit(void);
 int i6c_hal_init(void);
 
 void i6c_audio_deinit(void);
-int i6c_audio_init(short samplerate);
+int i6c_audio_init(int samplerate);
 void *i6c_audio_thread(void);
 
 int i6c_channel_bind(char index, char framerate);
